# server/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "StayConnected API",
    "description": "Distributed Active Climate Sensor Network"
  },
  "host": "localhost:5000",
  "basePath": "/api",
  "tags": [
    {
      "name": "Users",
      "description": "API for users in the system"
    },
    {
      "name": "Sites",
      "description": "API for sites in the system"
    },
    {
      "name": "Locations",
      "description": "API for locations in the system"
    },
    {
      "name": "Zones",
      "description": "API for zones in the system"
    },
    {
      "name": "Nodes",
      "description": "API for nodes in the system"
    },
    {
      "name": "Permissions",
      "description": "API for sites in the system"
    },
    {
      "name": "NoaaStations",
      "description": "API for sites in the system"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "Create new user in system",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User that we want to create",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New user is created",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get current user details",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/users/{userId}": {
      "parameters": [
        {
          "name": "userId",
          "in": "path",
          "required": true,
          "description": "ID of user that we want to find",
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get user with given ID",
        "responses": {
          "200": {
            "description": "User is found",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      },
      "delete": {
        "summary": "Delete user with given ID",
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "User is deleted",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      },
      "put": {
        "summary": "Update user with give ID",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User with new values of properties",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User is updated",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/permissions": {
      "post": {
        "tags": [
          "Permissions"
        ],
        "description": "Create new permission in system",
        "parameters": [
          {
            "name": "permission",
            "in": "body",
            "description": "Permission that we want to create",
            "schema": {
              "$ref": "#/definitions/Permission"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New permission is created",
            "schema": {
              "$ref": "#/definitions/Permission"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "required": [
        "user_id",
        "first_name",
        "last_name",
        "title",
        "role_id"
      ],
      "properties": {
        "user_id": {
          "type": "string"
        },
        "last_name": {
          "type": "string"
        },
        "first_name": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "role_id": {
          "type": "integer"
        }
      }
    },
    "Permission": {
      "required": [
        "role_id",
        "permissions",
        "role_name"
      ],
      "properties": {
        "role_id": {
          "type": "string"
        },
        "permissions": {
          "type": "string"
        },
        "role_name": {
          "type": "string"
        }
      }
    }
  }
}